+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.1] - 2019-02-18
+
+### Fixed
+
+- `web3x-codegen` failed when ABI had no constructor.
+- `web3x-codegen` generated bad deploy for some types.
+
+### Changed
+
+- Refactor some transaction logic.
+
 ## [3.0.0] - 2019-02-11
 
 ### Added
@@ -139,6 +150,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release of Typescript port from web3.js.
 
+[3.0.1]: https://github.com/xf00f/web3x/compare/v3.0.0...v3.0.1
 [3.0.0]: https://github.com/xf00f/web3x/compare/v2.0.4...v3.0.0
 [2.0.4]: https://github.com/xf00f/web3x/compare/v2.0.3...v2.0.4
 [2.0.3]: https://github.com/xf00f/web3x/compare/v2.0.2...v2.0.3